requirements.txt
defusedxml==0.7.1
fonttools==4.56.0
fpdf2==2.8.2
pillow==11.1.0
svgwrite==1.4.3
